+++ src/lib/marketing/Container/Container.svelte
@@ -2,18 +2,22 @@
 	interface Props {
 		as?: string;
 		children?: import('svelte').Snippet;
+        max?: boolean;
 	}
 
-	let { as = 'div', children }: Props = $props();
+	let { as = 'div', children, max = false
+    }: Props = $props();
+
+    let width= max ? '1400px' : '1000px';
 </script>
 
-<svelte:element this={as} class="container">
+<svelte:element this={as} class="container" style:width={width}>
 	{@render children?.()}
 </svelte:element>
 
 <style>
 	.container {
-		width: 1000px;
+		/*width: 1000px;*/
 		max-width: 100%;
 		padding: 0 15px;
 		margin: auto;