+++ src/routes/test/+page.svelte
@@ -1,3 +1,4 @@
+<!--THIS PAGE IS ONLY FOR TESTING-->
 <script>
 import Base from '$lib/components/Base/Base.svelte';
 import Header from '$lib/marketing/Header/Header.svelte';
@@ -5,13 +6,18 @@ import logo from '../img/logo/hyvor.svg';
 import Button from '../../lib/components/Button/Button.svelte';
 import IconBoxArrowUpRight from '@hyvor/icons/IconBoxArrowUpRight';
 import IconGithub from '@hyvor/icons/IconGithub';
+import Footer from '$lib/marketing/Footer/Footer.svelte';
+import  FooterLinkList from '$lib/marketing/Footer/FooterLinkList.svelte';
 </script>
 
 <svelte:head>
     <title>Test Page - HDS</title>
 </svelte:head>
 
-<Header {logo} name="HYVOR" subName="Design System">
+<Header {logo} name="HYVOR" subName="Design System" max={true}>
+    {#snippet center()}
+        <!-- Center content can go here -->
+    {/snippet}
     {#snippet end()}
         <Button size="small" as="a" href="https://hyvor.com" variant="invisible">HYVOR</Button>
         <Button as="a" size="small" href="https://github.com/hyvor/design" target="_blank">
@@ -35,3 +41,54 @@ import IconGithub from '@hyvor/icons/IconGithub';
         </p>
     </div>
 </Base>
+
+<Footer email="talk.support@hyvor.com" max={true}>
+    {#snippet center()}
+        <div class="footer-wrap">
+            <div style="display:flex" class="footer">
+                <FooterLinkList title="HYVOR TALK">
+                    <a href="/console">Console</a>
+                    <a href="/pricing">Pricing</a>
+                    <a href="/docs">Docs</a>
+                </FooterLinkList>
+
+                <FooterLinkList title="LEGAL">
+                    <a href="/terms">Terms</a>
+                    <a href="/privacy">Privacy Policy</a>
+                    <a href="https://hyvor.com/compliance" target="_blank">Compliance</a>
+                </FooterLinkList>
+
+                <FooterLinkList title="HYVOR">
+                    <a href="https://hyvor.com" target="_blank">hyvor.com</a>
+                    <a href="https://hyvor.com/#letter" target="_blank">About</a>
+                    <a href="https://hyvor.com/security" target="_blank">Security</a>
+                    <a href="https://status.hyvor.com/" target="_blank">System status</a>
+                    <a href="https://hyvor.com/blog" target="_blank">Blog</a>
+                </FooterLinkList>
+
+                <FooterLinkList title="RESOURCES">
+                    <a href="/blog">Blog</a>
+                    <a href="/docs">Documentation</a>
+                    <a href="https://hyvor.design" target="_blank">Design System</a>
+                </FooterLinkList>
+            </div>
+        </div>
+    {/snippet}
+</Footer>
+
+<style lang="scss">
+    // media queries
+    @media (max-width: 768px) {
+        .footer-wrap {
+            padding: 20px;
+        }
+        .footer {
+            flex-direction: column;
+            gap: 20px;
+        }
+    }
+</style>
+
+
+
+