+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.1.21 - 2025-10-21
 - Added `Accordion` component for collapsible content sections
 - Added `x-large` size option to `Button` component
+- Added `max`prop to `Header` component to change width based on page type
 
 ## 1.1.19 - 2025-09-29